requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
django-graphene-permissions==0.0.4
django-graphql-jwt==0.3.4
django-redis==5.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2==2.9.1
PyJWT==2.2.0
pyotp==2.6.0
pytz==2021.3
redis==3.5.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
